% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_record_linkage.R
\name{active_comparison}
\alias{active_comparison}
\title{Active comparison between records}
\usage{
active_comparison(data)
}
\arguments{
\item{data}{dataframe with espn names and fifa names and possibly additional
information}
}
\value{
vector with all answers
}
\description{
This function shows one at the time all the comparisons,
for all unique espn players in data
}
