% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_record_linkage.R
\name{active_insert}
\alias{active_insert}
\title{Active insertion of fifa names}
\usage{
active_insert(data)
}
\arguments{
\item{data}{dataframe with espn names and possibly additional
information}
}
\value{
vector with all answers. Either NA or the answer of the user
}
\description{
This function allow the user to insert the fifa names for all unique espn players
in data
}
