% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_minus.R
\name{add_red_card_info}
\alias{add_red_card_info}
\title{Accounting for red cards}
\usage{
add_red_card_info(game_commentary, game_lineup, type_card = "red")
}
\arguments{
\item{game_commentary}{from fcscrapR}

\item{game_lineup}{computed from the package}

\item{type_card}{is the type of card to consider. It has to be red,
left it flexible for testing purposes}
}
\description{
This function adds the info about red card to the game_lineup
}
