% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_record_linkage.R
\name{block_comparison}
\alias{block_comparison}
\title{Block comparison}
\usage{
block_comparison(records)
}
\arguments{
\item{records}{table of records with espn, fifa name and some additional info}
}
\value{
the user answer
}
\description{
This function shows  the espn name with one or more fifa candidates and ask the
user to decide which one is the most plausible match. The user can input
the row number of the selected match, or 'y' if the only row is shown and the
user believes that is a match, 0 if the user doesn't see any plausible match
}
