% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sofifa.R
\name{clean_sofifa_table}
\alias{clean_sofifa_table}
\title{Clean raw sofifa table}
\usage{
clean_sofifa_table(tab)
}
\arguments{
\item{tab}{data-frame of sofifa table extracted from \code{scrape_sofifa_table}}
}
\description{
Clean raw sofifa table
}
