% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_minus.R
\name{create_design_matrix}
\alias{create_design_matrix}
\title{Real Plus Minus matrix}
\usage{
create_design_matrix(lineups, segments)
}
\arguments{
\item{lineups}{all lineups by league stacked in a matrix}

\item{segments}{all segments by league stacked in a matrix}
}
\description{
This function creates the design matrix for a given league, given the lineups
and the segmentation matrix
}
