% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_minus.R
\name{create_segmentation}
\alias{create_segmentation}
\title{Creating segmentation}
\usage{
create_segmentation(game_lineup)
}
\arguments{
\item{game_lineup}{computed from the package}

\item{game_commentary}{from fcscrapR}
}
\description{
This function creates segmentation every time a the configuration
on the pitch changes
}
