% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_lineups.R
\name{get_player_name}
\alias{get_player_name}
\title{Extract the player name given a string with player data info}
\usage{
get_player_name(player_by_player_string)
}
\arguments{
\item{player_by_player_string}{string with a single player info}
}
\description{
This function extracts the player name given a string with player data info
}
