% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_minus.R
\name{get_rpm}
\alias{get_rpm}
\title{Compute the real-plus minus}
\usage{
get_rpm(design_matrix, threshold = 0, minutes, response = "goal",
  type = "both")
}
\arguments{
\item{design_matrix}{0, -1, 1 matrix with info about shots and goals}

\item{threshold}{for minutes, by default 0}

\item{minutes}{dataframe with minutes info by player}

\item{response}{character to chose as response, either goal, or shots}

\item{type}{can be offensive, defensive or both which takes the differnce
in the response}
}
\description{
This function computes real plus minus and threshold by minutes player
if requested
}
