% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_minus.R
\name{get_shot_attempt_by_segment}
\alias{get_shot_attempt_by_segment}
\title{Creating the shot attempt by segmentation for both teams}
\usage{
get_shot_attempt_by_segment(game_commentary, segmentation_matrix)
}
\arguments{
\item{game_commentary}{from fcscrapR}

\item{segmentation_matrix}{a matrix with each row a segment, each coloumn
the start and the end of the segment}
}
\description{
This function creates the shot attempt matrix by segmentation for both teams
}
