% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_record_linkage.R
\name{insert_fifa_name}
\alias{insert_fifa_name}
\title{Insert fifa name}
\usage{
insert_fifa_name(record)
}
\arguments{
\item{record}{is the espn name with possibly additional info on the player}
}
\value{
the answer of the user
}
\description{
This function allows the user to insert the fifa name, given the espn name
and some other info
}
