% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_lineups.R
\name{scrape_lineup}
\alias{scrape_lineup}
\title{Scrape lineups for ESPN Game ID}
\usage{
scrape_lineup(game_id)
}
\arguments{
\item{game_id}{espn id}
}
\value{
list with the two line ups plus subs info
}
\description{
This function scrapes lineup for a given game on espn
}
