% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sofifa.R
\name{scrape_sofifa_table}
\alias{scrape_sofifa_table}
\title{Scrape sofifa tables}
\usage{
scrape_sofifa_table(url)
}
\arguments{
\item{url}{url for the sofifa table}
}
\description{
Scrape sofifa tables
}
