% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_lineups.R
\name{split_lineup_when_no_stats}
\alias{split_lineup_when_no_stats}
\title{Split the lineup when no stats are available for some players}
\usage{
split_lineup_when_no_stats(team_lineup)
}
\arguments{
\item{team_lineup}{team lineup extracted by espn removed of buzzwords and extra
spaces}
}
\description{
This function splits the lineup info when the stats are missing for some players
}
